+++ Content.Server/Temperature/Systems/TemperatureSystem.cs
@@ -39,14 +39,14 @@ public sealed class TemperatureSystem : SharedTemperatureSystem
 
     public override void Initialize()
     {
+        base.Initialize();
+
         SubscribeLocalEvent<TemperatureComponent, OnTemperatureChangeEvent>(EnqueueDamage);
         SubscribeLocalEvent<TemperatureComponent, AtmosExposedUpdateEvent>(OnAtmosExposedUpdate);
         SubscribeLocalEvent<TemperatureComponent, RejuvenateEvent>(OnRejuvenate);
         SubscribeLocalEvent<AlertsComponent, OnTemperatureChangeEvent>(ServerAlert);
         SubscribeLocalEvent<TemperatureProtectionComponent, InventoryRelayedEvent<ModifyChangedTemperatureEvent>>(
             OnTemperatureChangeAttempt);
-
-        SubscribeLocalEvent<InternalTemperatureComponent, MapInitEvent>(OnInit);
 
         // Allows overriding thresholds based on the parent's thresholds.
         SubscribeLocalEvent<TemperatureComponent, EntParentChangedMessage>(OnParentChange);
@@ -56,37 +56,7 @@ public sealed class TemperatureSystem : SharedTemperatureSystem
             OnParentThresholdShutdown);
     }
 
-    public override void Update(float frameTime)
-    {
-        base.Update(frameTime);
-
-        // conduct heat from the surface to the inside of entities with internal temperatures
-        var query = EntityQueryEnumerator<InternalTemperatureComponent, TemperatureComponent>();
-        while (query.MoveNext(out var uid, out var comp, out var temp))
-        {
-            // don't do anything if they equalised
-            var diff = Math.Abs(temp.CurrentTemperature - comp.Temperature);
-            if (diff < 0.1f)
-                continue;
-
-            // heat flow in W/m^2 as per fourier's law in 1D.
-            var q = comp.Conductivity * diff / comp.Thickness;
-
-            // convert to J then K
-            var joules = q * comp.Area * frameTime;
-            var degrees = joules / GetHeatCapacity(uid, temp);
-            if (temp.CurrentTemperature < comp.Temperature)
-                degrees *= -1;
-
-            // exchange heat between inside and surface
-            comp.Temperature += degrees;
-            ForceChangeTemperature(uid, temp.CurrentTemperature - degrees, temp);
-        }
-
-        UpdateDamage(frameTime);
-    }
-
-    private void UpdateDamage(float frameTime)
+    protected override void UpdateDamage(float frameTime)
     {
         _accumulatedFrametime += frameTime;
 
@@ -125,14 +95,6 @@ public sealed class TemperatureSystem : SharedTemperatureSystem
         var heat = temperatureDelta * (airHeatCapacity * heatCapacity /
                                        (airHeatCapacity + heatCapacity));
         ChangeHeat(uid, heat * temperature.AtmosTemperatureTransferEfficiency, temperature: temperature);
-    }
-
-    private void OnInit(EntityUid uid, InternalTemperatureComponent comp, MapInitEvent args)
-    {
-        if (!TryComp<TemperatureComponent>(uid, out var temp))
-            return;
-
-        comp.Temperature = temp.CurrentTemperature;
     }
 
     private void OnRejuvenate(EntityUid uid, TemperatureComponent comp, RejuvenateEvent args)
